network call logic/ bugs still remain
