+++ app/src/main/java/com/example/c4q/capstone/LoginActivity.java
@@ -17,7 +17,6 @@ import android.view.View;
 
 import com.example.c4q.capstone.database.model.publicuserdata.PublicUser;
 import com.example.c4q.capstone.database.model.publicuserdata.UserSearch;
-import com.example.c4q.capstone.network.barzz.NetworkCall;
 import com.example.c4q.capstone.userinterface.user.EditProfileActivity;
 import com.example.c4q.capstone.userinterface.user.SettingsActivity;
 import com.example.c4q.capstone.userinterface.user.UserProfileActivity;
@@ -71,7 +70,7 @@ public class LoginActivity extends AppCompatActivity {
          * */
 
         getUserData();
-      //  NetworkCall.start("10001");
+      //  BarzzNetworkCall.start("10001");
 
         List<AuthUI.IdpConfig> providers = Arrays.asList(
                 new AuthUI.IdpConfig.Builder(AuthUI.EMAIL_PROVIDER).build(),
